[REFACTOR] #2 - package 명 변경
