+++ app/src/main/AndroidManifest.xml
@@ -14,10 +14,10 @@
         android:theme="@style/Theme.Android_repo_05"
         tools:targetApi="31">
         <activity
-            android:name=".ui.SearchActivity"
+            android:name=".ui.activities.SearchActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.ProfileActivity"
+            android:name=".ui.activities.ProfileActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
