+++ app/src/main/res/layout/activity_main.xml
@@ -42,7 +42,7 @@
             android:layout_marginHorizontal="24dp"
             app:layout_constraintTop_toBottomOf="@id/l_tb_main">
 
-            <com.example.android_repo_05.customview.MainTabButton
+            <com.example.android_repo_05.customviews.MainTabButton
                 android:id="@+id/tab_btn_issue"
                 android:layout_width="0dp"
                 android:layout_height="match_parent"
@@ -52,7 +52,7 @@
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintEnd_toStartOf="@id/tab_btn_notification"/>
 
-            <com.example.android_repo_05.customview.MainTabButton
+            <com.example.android_repo_05.customviews.MainTabButton
                 android:id="@+id/tab_btn_notification"
                 android:layout_width="0dp"
                 android:layout_height="match_parent"
