+++ app/src/main/res/layout/activity_search.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.SearchActivity">
+    tools:context=".ui.activities.SearchActivity">
 
     <TextView
         android:id="@+id/textView4"
